+++ cookbooks/cwb-chef-server/berks-cookbooks/chef-ingredient/libraries/chef_ingredient_resource.rb
@@ -1,6 +1,6 @@
 #
 # Author:: Joshua Timberman <joshua@chef.io
-# Copyright (c) 2015, Chef Software, Inc. <legal@chef.io>
+# Copyright (c) 2015-2016, Chef Software, Inc. <legal@chef.io>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,6 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
 class Chef
   class Resource
     class ChefIngredient < Chef::Resource::LWRPBase
@@ -21,23 +22,35 @@ class Chef
 
       actions :install, :uninstall, :remove, :reconfigure, :upgrade
       default_action :install
-      state_attrs :installed
 
       attribute :product_name, kind_of: String, name_attribute: true
-      attribute :installed, kind_of: [TrueClass, FalseClass, NilClass], default: false
-      attribute :reconfigure, kind_of: [TrueClass, FalseClass], default: false
-      attribute :config, kind_of: String, default: nil
+      attribute :config, kind_of: String
+
+      # Attributes for determining what version to install from which channel
+      attribute :version, kind_of: [String, Symbol], default: :latest
+      attribute :channel, kind_of: Symbol, default: :stable, equal_to: [:stable, :current, :unstable]
 
       # Attribute to install package from local file
-      attribute :package_source, kind_of: String, default: nil
+      attribute :package_source, kind_of: String
 
-      # Attributes for reconfigure step
+      # Sets the *-ctl command to use when doing reconfigure
       attribute :ctl_command, kind_of: String
 
-      # Attributes for package
+      # Attributes for package resources used on rhel and debian platforms
       attribute :options, kind_of: String
-      attribute :version, kind_of: [String, Symbol], default: :latest
-      attribute :timeout, kind_of: [Integer, String, NilClass], default: nil
+      attribute :timeout, kind_of: [Integer, String, NilClass]
+
+      # Attribute to accept the license when applicable
+      attribute :accept_license, kind_of: [TrueClass, FalseClass], default: false
+
+      # Attribute to enable selecting packages built for earlier versions in
+      # platforms that are not yet officially added to Chef support matrix
+      attribute :platform_version_compatibility_mode, kind_of: [TrueClass, FalseClass], default: false
+
+      # Atttibutes for configuring a specific platform package
+      attribute :platform, kind_of: String
+      attribute :platform_version, kind_of: String
+      attribute :architecture, kind_of: String
     end
   end
 end