+++ cookbooks/cwb-chef-server/berks-cookbooks/chef-ingredient/libraries/helpers.rb
@@ -1,6 +1,6 @@
 #
 # Author:: Joshua Timberman <joshua@chef.io>
-# Copyright (c) 2014-2015, Chef Software, Inc. <legal@chef.io>
+# Copyright (c) 2014-2016, Chef Software, Inc. <legal@chef.io>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,61 +17,184 @@
 
 module ChefIngredientCookbook
   module Helpers
-    def chef_ctl_command(product)
-      if new_resource.respond_to?(:version)
-        product_lookup(product, version_string(new_resource.version))['ctl-command']
-      else
-        product_lookup(product)['ctl-command']
-      end
+    ########################################################################
+    # Product details lookup helpers
+    ########################################################################
+
+    #
+    # Returns the ctl-command to be used for omnibus_service resource.
+    # Notice that we do not include version in our lookup since omnibus_service
+    # is not aware of it.
+    #
+    def ctl_command_for_product(product)
+      ensure_mixlib_install_gem_installed!
+
+      PRODUCT_MATRIX.lookup(product).ctl_command
     end
 
-    def version_string(vers)
-      return '0.0.0' if vers.to_sym == :latest
-      vers
-    end
-
+    #
+    # Returns the package name for a given product and version.
+    #
     def ingredient_package_name
-      product_lookup(new_resource.product_name, version_string(new_resource.version))['package-name']
+      ensure_mixlib_install_gem_installed!
+
+      PRODUCT_MATRIX.lookup(new_resource.product_name, new_resource.version).package_name
     end
 
-    def local_package_resource
-      return :dpkg_package if node['platform_family'] == 'debian'
-      return :rpm_package  if node['platform_family'] == 'rhel'
-      :package # fallback if there's no platform match
+    #
+    # Returns the ctl-command for a chef_ingredient resource
+    #
+    def ingredient_ctl_command
+      ensure_mixlib_install_gem_installed!
+
+      new_resource.ctl_command || PRODUCT_MATRIX.lookup(new_resource.product_name, new_resource.version).ctl_command
     end
 
-    def package_resource(ingredient_action)
-      presource = new_resource.package_source.nil? ? :package : local_package_resource
+    #
+    # Returns the ctl-command for a chef_ingredient resource
+    #
+    def ingredient_config_file(product_name)
+      ensure_mixlib_install_gem_installed!
 
-      declare_resource presource, new_resource.product_name do
-        package_name ingredient_package_name
-        options new_resource.options
-        version install_version if Mixlib::Versioning.parse(version_string(new_resource.version)) > '0.0.0'
-        source new_resource.package_source
-        timeout new_resource.timeout
-        action ingredient_action
+      PRODUCT_MATRIX.lookup(product_name).config_file
+    end
+
+    ########################################################################
+    # Helpers installing gem prerequisites.
+    ########################################################################
+
+    #
+    # Ensures mixlib-versioning gem is installed and loaded.
+    #
+    def ensure_mixlib_versioning_gem_installed!
+      node.run_state[:mixlib_versioning_gem_installed] ||= begin # ~FC001
+        install_gem_from_rubygems('mixlib-versioning', '~> 1.1')
+
+        require 'mixlib/versioning'
+        true
       end
     end
 
-    def install_mixlib_versioning
-      # We need Mixlib::Versioning in the library helpers for
-      # parsing the version string.
-      chef_gem "#{new_resource.product_name}-mixlib-versioning" do # ~FC009 foodcritic needs an update
-        package_name 'mixlib-versioning'
-        compile_time true
+    #
+    # Ensures mixlib-install gem is installed and loaded.
+    #
+    def ensure_mixlib_install_gem_installed!
+      node.run_state[:mixlib_install_gem_installed] ||= begin # ~FC001
+        if node['chef-ingredient']['mixlib-install']['git_ref']
+          install_gem_from_source(
+            'https://github.com/chef/mixlib-install.git',
+            node['chef-ingredient']['mixlib-install']['git_ref'],
+            'mixlib-install'
+          )
+        else
+          install_gem_from_rubygems('mixlib-install', ['~> 2.1', '>= 2.1.12'])
+        end
+
+        require 'mixlib/install'
+        require 'mixlib/install/product'
+        true
       end
-
-      require 'mixlib/versioning'
     end
 
-    def create_repository
-      cleanup_old_repo_config if ::File.exist?(old_ingredient_repo_file)
-      include_recipe "#{package_repo_type}-chef" if new_resource.package_source.nil?
+    #
+    # Helper method to install a gem from rubygems at compile time.
+    #
+    def install_gem_from_rubygems(gem_name, gem_version)
+      Chef::Log.debug("Installing #{gem_name} v#{gem_version} from Rubygems.org")
+      chefgem = Chef::Resource::ChefGem.new(gem_name, run_context)
+      chefgem.version(gem_version)
+      chefgem.run_action(:install)
     end
 
-    def package_repo_type
-      return 'apt' if node['platform_family'] == 'debian'
-      return 'yum' if node['platform_family'] == 'rhel'
+    #
+    # Helper method to install a gem from source at compile time.
+    #
+    def install_gem_from_source(repo_url, git_ref, gem_name = nil)
+      uri = URI.parse(repo_url)
+      repo_basename = ::File.basename(uri.path)
+      repo_name = repo_basename.match(/(?<name>.*)\.git/)[:name]
+      gem_name ||= repo_name
+
+      Chef::Log.debug("Building #{gem_name} gem from source")
+
+      gem_clone_path = ::File.join(Chef::Config[:file_cache_path], repo_name)
+      gem_file_path  = ::File.join(gem_clone_path, "#{gem_name}-*.gem")
+
+      checkout_gem = Chef::Resource::Git.new(gem_clone_path, run_context)
+      checkout_gem.repository(repo_url)
+      checkout_gem.revision(git_ref)
+      checkout_gem.run_action(:sync)
+
+      ::FileUtils.rm_rf gem_file_path
+
+      build_gem = Chef::Resource::Execute.new("build-#{gem_name}-gem", run_context)
+      build_gem.cwd(gem_clone_path)
+      build_gem.command(
+        <<-EOH
+#{::File.join(RbConfig::CONFIG['bindir'], 'gem')} build #{gem_name}.gemspec
+        EOH
+      )
+      build_gem.run_action(:run) if checkout_gem.updated?
+
+      install_gem = Chef::Resource::ChefGem.new(gem_name, run_context)
+      install_gem_file_path = if windows?
+                                Dir.glob(gem_file_path.tr('\\', '/')).first
+                              else
+                                Dir.glob(gem_file_path).first
+                              end
+      install_gem.source(install_gem_file_path)
+      install_gem.run_action(:install) if build_gem.updated?
+    end
+
+    ########################################################################
+    # Version helpers
+    ########################################################################
+
+    #
+    # Returns if a given version is equivalent to :latest
+    #
+    def version_latest?(vers)
+      vers == :latest || vers == '0.0.0' || vers == 'latest'
+    end
+
+    #
+    # Returns true if the custom repo recipe was specified
+    #
+    def use_custom_repo_recipe?
+      node['chef-ingredient'].attribute?('custom-repo-recipe')
+    end
+
+    #
+    # Returns the custom setup recipe name
+    #
+    # When the user specifies this attribute chef-ingredient will not configure
+    # our default packagecloud Chef repositories and instead it will include the
+    # custom recipe. This will eliminate the hard dependency to the internet.
+    #
+    def custom_repo_recipe
+      node['chef-ingredient']['custom-repo-recipe']
+    end
+
+    #
+    # Returns the version string to use in package resource for all platforms.
+    #
+    def version_for_package_resource
+      ensure_mixlib_versioning_gem_installed!
+
+      version_string = if version_latest?(new_resource.version)
+                         '0.0.0'
+                       else
+                         new_resource.version
+                       end
+
+      v = Mixlib::Versioning.parse(version_string)
+      version = "#{v.major}.#{v.minor}.#{v.patch}"
+      version << "~#{v.prerelease}" if v.prerelease? && !v.prerelease.match(/^\d$/)
+      version << "+#{v.build}" if v.build?
+      version << '-1' unless version =~ /-1$/
+      version << rhel_append_version if node['platform_family'] == 'rhel' &&
+                                        !version.match(/#{rhel_append_version}$/)
+      version
     end
 
     def rhel_major_version
@@ -79,183 +202,45 @@ module ChefIngredientCookbook
       node['platform_version']
     end
 
-    def install_version
-      require 'mixlib/versioning'
-      v = Mixlib::Versioning.parse(version_string(new_resource.version))
-      version = "#{v.major}.#{v.minor}.#{v.patch}"
-      version << "~#{v.prerelease}" if v.prerelease? && !v.prerelease.match(/^\d$/)
-      version << "+#{v.build}" if v.build?
-      version << '-1' unless version.match(/-1$/)
-      version << rhel_append_version if node['platform_family'] == 'rhel' &&
-                                        !version.match(/#{rhel_append_version}$/)
-      version
-    end
-
     def rhel_append_version
       ".el#{rhel_major_version}"
     end
 
-    def old_ingredient_repo_file
-      return '/etc/apt/sources.list.d/chef_stable_.list' if node['platform_family'] == 'debian'
-      return '/etc/yum.repos.d/chef_stable_.repo' if node['platform_family'] == 'rhel'
-    end
+    ########################################################################
+    # ingredient_config helpers
+    ########################################################################
 
-    def cleanup_old_repo_config
-      file old_ingredient_repo_file do
-        action :delete
-      end
-    end
-
-    def ctl_command
-      new_resource.ctl_command || chef_ctl_command(new_resource.product_name)
-    end
-
-    def reconfigure
-      ctl_cmd = ctl_command
-      execute "#{new_resource.product_name}-reconfigure" do
-        command "#{ctl_cmd} reconfigure"
-      end
-    end
-
-    # When updating this, also update PRODUCT_MATRIX.md
-    def product_matrix
-      {
-        'analytics' => {
-          'package-name' => 'opscode-analytics',
-          'ctl-command'  => 'opscode-analytics-ctl',
-          'config-file'  => '/etc/opscode-analytics/opscode-analytics.rb'
-        },
-        'chef' => {
-          'package-name' => 'chef',
-          'ctl-command'  => nil,
-          'config-file'  => nil
-        },
-        'chef-ha' => {
-          'package-name' => 'chef-ha',
-          'ctl-command'  => nil,
-          'config-file'  => '/etc/opscode/chef-server.rb'
-        },
-        'chef-marketplace' => {
-          'package-name' => 'chef-marketplace',
-          'ctl-command'  => 'chef-marketplace-ctl',
-          'config-file'  => '/etc/chef-marketplace/marketplace.rb'
-        },
-        'chef-server' => {
-          'package-name' => 'chef-server-core',
-          'ctl-command'  => 'chef-server-ctl',
-          'config-file'  => '/etc/opscode/chef-server.rb'
-        },
-        'chef-sync' => {
-          'package-name' => 'chef-sync',
-          'ctl-command'  => 'chef-sync-ctl',
-          'config-file'  => '/etc/chef-sync/chef-sync.rb'
-        },
-        'chefdk' => {
-          'package-name' => 'chefdk',
-          'ctl-command'  => nil,
-          'config-file'  => nil
-        },
-        'delivery' => {
-          'package-name' => 'delivery',
-          'ctl-command'  => 'delivery-ctl',
-          'config-file'  => '/etc/delivery/delivery.rb'
-        },
-        'delivery-cli' => {
-          'package-name' => 'delivery-cli',
-          'ctl-command'  => nil,
-          'config-file'  => nil
-        },
-        'manage' => {
-          'package-name' => 'chef-manage',
-          'ctl-command'  => 'chef-manage-ctl',
-          'config-file'  => '/etc/opscode-manage/manage.rb'
-        },
-        'private-chef' => {
-          'package-name' => 'private-chef',
-          'ctl-command'  => 'private-chef-ctl',
-          'config-file'  => '/etc/opscode/private-chef.rb'
-        },
-        'push-client' => {
-          'package-name' => 'chef-push-client',
-          'ctl-command'  => nil,
-          'config-file'  => nil
-        },
-        'push-server' => {
-          'package-name' => 'chef-push-server',
-          'ctl-command'  => 'chef-push-ctl',
-          'config-file'  => '/etc/opscode-push-jobs-server/opscode-push-jobs-server.rb'
-        },
-        'reporting' => {
-          'package-name' => 'opscode-reporting',
-          'ctl-command'  => 'opscode-reporting-ctl',
-          'config-file'  => '/etc/opscode-reporting/opscode-reporting.rb'
-        },
-        'supermarket' => {
-          'package-name' => 'supermarket',
-          'ctl-command'  => 'supermarket-ctl',
-          'config-file'  => '/etc/supermarket/supermarket.json'
-        }
-      }
-    end
-
-    # Version has a default value of 0.0.0 so that it is a valid
-    # string for the Mixlib::Versioning.parse method. This implies
-    # "latest", but "latest" is not a value that is valid for
-    # mixlib/versioning.
-    def product_lookup(product, version = '0.0.0')
-      unless product_matrix.key?(product)
-        Chef::Log.fatal("We don't have a product, '#{product}'. Please specify a valid product name:")
-        Chef::Log.fatal(product_matrix.keys.join(' '))
-        fail
-      end
-
-      require 'mixlib/versioning'
-      v = Mixlib::Versioning.parse(version_string(version))
-
-      data = product_matrix[product]
-
-      # We want to validate that we're getting a version that is valid
-      # for the Chef Server. However, since the default is 0.0.0,
-      # implying latest, we need to additionally ensure that the
-      # bottom version is something valid. If we don't have the check
-      # in the elsif, it will say that 0.0.0 is not a valid version.
-      if (product == 'chef-server')
-        if (v < Mixlib::Versioning.parse('12.0.0')) && (v > Mixlib::Versioning.parse('11.0.0'))
-          data['package-name'] = 'chef-server'
-        elsif (v < Mixlib::Versioning.parse('11.0.0')) && (v > Mixlib::Versioning.parse('1.0.0'))
-          Chef::Log.fatal("Invalid version specified, '#{version}' for #{product}!")
-          fail
-        end
-      elsif (product == 'manage') && (v < Mixlib::Versioning.parse('2.0.0'))
-        data['package-name'] = 'opscode-manage'
-        data['ctl-command'] = 'opscode-manage-ctl'
-      # TODO: When Chef Push server and client 2.0 are released, we
-      # need to implement similar logic to chef-server, so that the
-      # default "latest" version, 0.0.0 (no constraint) doesn't result
-      # in the old package.
-      elsif (product == 'push-server') && (v < Mixlib::Versioning.parse('2.0.0'))
-        data['package-name'] = 'opscode-push-jobs-server'
-        data['ctl-command'] = 'opscode-push-jobs-server-ctl'
-      elsif (product == 'push-client') && (v < Mixlib::Versioning.parse('2.0.0'))
-        data['package-name'] = 'opscode-push-jobs-client'
-      end
-
-      data
-    end
-
+    #
+    # Adds given config information for the given product to the run_state so
+    # that it can be retrieved later.
+    #
     def add_config(product, content)
       return if content.nil?
 
-      node.run_state[:ingredient_config_data] ||= {}
-      node.run_state[:ingredient_config_data][product] ||= ''
-      node.run_state[:ingredient_config_data][product] += content
+      # FC001: Use strings in preference to symbols to access node attributes
+      # foodcritic thinks we are accessing a node attribute
+      node.run_state[:ingredient_config_data] ||= {}              # ~FC001
+      node.run_state[:ingredient_config_data][product] ||= ''     # ~FC001
+      node.run_state[:ingredient_config_data][product] += content unless node.run_state[:ingredient_config_data][product].include?(content) # ~FC001
     end
 
+    #
+    # Returns the collected config information for the given product.
+    #
     def get_config(product)
-      node.run_state[:ingredient_config_data] ||= {}
-      node.run_state[:ingredient_config_data][product] ||= ''
+      # FC001: Use strings in preference to symbols to access node attributes
+      # foodcritic thinks we are accessing a node attribute
+      node.run_state[:ingredient_config_data] ||= {}          # ~FC001
+      node.run_state[:ingredient_config_data][product] ||= '' # ~FC001
     end
 
+    ########################################################################
+    # misc helpers
+    ########################################################################
+
+    #
+    # Returns true if a given fqdn resolves, false otherwise.
+    #
     def fqdn_resolves?(fqdn)
       require 'resolv'
       Resolv.getaddress(fqdn)
@@ -263,14 +248,110 @@ module ChefIngredientCookbook
     rescue Resolv::ResolvError, Resolv::ResolvTimeout
       false
     end
-
     module_function :fqdn_resolves?
-  end
-end
 
-module ChefServerIngredientsCookbook
-  module Helpers
-    include ChefIngredientCookbook::Helpers
-    alias_method :chef_server_ctl_command, :chef_ctl_command
+    #
+    # Declares a resource that will fail the chef run when signalled.
+    #
+    def declare_chef_run_stop_resource
+      # We do not supply an option to turn off stopping the chef client run
+      # after a version change. As the gems shipped with omnitruck artifacts
+      # change, chef-client runs *WILL* occasionally break on minor version
+      # updates of chef, so we *MUST* stop the chef-client run when its version
+      # changes. The gems versions that chef-client started with will not
+      # necessarily exist after an upgrade.
+      ruby_block 'stop chef run' do
+        action :nothing
+        block do
+          Chef::Application.fatal! 'Chef version has changed during the run. Stopping the current Chef run. Please run chef again.'
+        end
+      end
+    end
+
+    def windows?
+      node['platform_family'] == 'windows'
+    end
+
+    #
+    # Creates a Mixlib::Install instance using the common attributes of
+    # chef_ingredient resource that can be used for querying builds or
+    # generating install scripts.
+    #
+    def installer
+      @installer ||= begin
+        ensure_mixlib_install_gem_installed!
+
+        options = {
+          product_name: new_resource.product_name,
+          channel: new_resource.channel,
+          product_version: new_resource.version,
+          platform_version_compatibility_mode: new_resource.platform_version_compatibility_mode,
+        }.tap do |opt|
+          opt[:shell_type] = :ps1 if windows?
+        end
+
+        platform_details = Mixlib::Install.detect_platform
+
+        platform_details.tap do |opt|
+          opt[:platform] = new_resource.platform if new_resource.platform
+          opt[:platform_version] = new_resource.platform_version if new_resource.platform_version
+          opt[:architecture] = new_resource.architecture if new_resource.architecture
+        end
+
+        options.merge!(platform_details)
+
+        Mixlib::Install.new(options)
+      end
+    end
+
+    #
+    # Returns package installer options with any required
+    # options based on platform
+    #
+    def package_options_with_force
+      options = new_resource.options
+
+      # Ubuntu 10.10 and Debian 6 require the `--force-yes` option
+      # for package installs
+      if (platform?('ubuntu') && node['platform_version'] == '10.04') ||
+         (platform?('debian') && node['platform_version'].start_with?('6'))
+        if options.nil?
+          options = '--force-yes'
+        else
+          options << ' --force-yes'
+        end
+      end
+
+      options
+    end
+
+    #
+    # Checks the deprecated properties of chef-ingredient and prints warning
+    # messages if any of them are being used.
+    #
+    def check_deprecated_properties
+      # Historically we have had chef- and opscode- in front of most of our
+      # packages. But with our move to bintray we have standardized on names
+      # without any prefixes except some products.
+      if !%w(chef-backend chef-server chef-server-ha-provisioning).include?(new_resource.product_name) &&
+         (match = new_resource.product_name.match(/(chef-|opscode-)(?<product_key>.*)/))
+
+        new_product_key = match[:product_key]
+        Chef::Log.warn "product_name '#{new_resource.product_name}' is deprecated and it will be removed in the future versions of chef-ingredient. Use '#{new_product_key}' instead of '#{new_resource.product_name}'."
+        new_resource.product_name(new_product_key)
+      else
+        # We also have a specific case we need to handle for push-client and push-server
+        deprecated_product_names = {
+          'push-client' => 'push-jobs-client',
+          'push-server' => 'push-jobs-server',
+        }
+
+        if deprecated_product_names.keys.include?(new_resource.product_name)
+          new_product_key = deprecated_product_names[new_resource.product_name]
+          Chef::Log.warn "product_name '#{new_resource.product_name}' is deprecated and it will be removed in the future versions of chef-ingredient. Use '#{new_product_key}' instead of '#{new_resource.product_name}'."
+          new_resource.product_name(new_product_key)
+        end
+      end
+    end
   end
 end