+++ cookbooks/cwb-chef-server/berks-cookbooks/chef-ingredient/recipes/default.rb
@@ -1,14 +1,11 @@
 #
-# Cookbook:: chef-server
-# Recipe:: addons
-#
-# Copyright:: 2015-2016, Chef Software, Inc.
+# Copyright:: 2015-2016, Chef Software, Inc. <legal@chef.io>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#    http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,10 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-node['chef-server']['addons'].each do |addon, ver|
-  chef_ingredient addon do
-    accept_license node['chef-server']['accept_license'] unless node['chef-server']['accept_license'].nil?
-    notifies :reconfigure, "chef_ingredient[#{addon}]"
-    version ver if defined?(ver)
-  end
-end
+
+# This recipe is included to make our tests happy.
+# By intention, it does not contain anything.