+++ cookbooks/cwb-chef-server/berks-cookbooks/chef-server/recipes/default.rb
@@ -1,5 +1,8 @@
 #
-# Copyright:: Copyright (c) 2012-2015 Chef Software, Inc.
+# Cookbook:: chef-server
+# Recipe:: default
+#
+# Copyright:: 2015-2016, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -25,7 +28,7 @@ end
 
 chef_ingredient 'chef-server' do
   extend ChefServerCookbook::Helpers
-  version node['chef-server']['version']
+  version node['chef-server']['version'] unless node['chef-server']['version'].nil?
   package_source node['chef-server']['package_source']
   config <<-EOS
 topology "#{node['chef-server']['topology']}"