+++ cookbooks/cwb-chef-server/berks-cookbooks/chef-ingredient/CHANGELOG.md
@@ -1,3 +1,121 @@
+# chef-ingredient Cookbook CHANGELOG
+This file is used to list changes made in each version of the chef-ingredient cookbook.
+
+## 1.1.0 (2017-03-01)
+
+- Test with local delivery and not Rake
+- Remove sensitive property for Chef 13 compatibility as this properly is provided by chef-client now for us by any resource and doesn't need to be defined
+- Test in Travis CI with kitchen-dokken and convert tests to InSpec
+
+## 1.0.1 (2017-02-22)
+
+- Testing cleanup for Chef 13 compatibility and testing on the latest platforms
+
+## 1.0.0 (2017-02-15)
+
+- Require Chef 12.5+ and remove compat_resource dependency
+- Use mixlib-install >= 2.1.12 - this brings in an important fix for the `delivery` -> `automate` package rename. See the [Discourse announcement](https://discourse.chef.io/t/chef-automate-install-package-renaming-in-0-7-14-available/10429/1) for details on the rename
+
+## 0.21.4 (2017-02-13)
+- Add properties to override the platform details of a `chef_ingredient` product to install
+
+## 0.21.3 (2017-02-02)
+- Add timeout to package resource created by configure_from_source_package
+
+## 0.21.2 (2016-10-26)
+- Fix issue when failed package installs using OmnitruckHandler would not raise a converge error on subsequent runs
+
+## 0.21.1 (2016-10-25)
+- Update SUSE platform to use DefaultHandler
+
+## 0.21.0 (2016-09-26)
+- Update mixlib-install to version 2.0 (PackageRouter support)
+
+## 0.20.0 (2016-09-08)
+- Remove extraneous converge_by that caused downloads to show as converged on every run
+- Use compat_resource cookbook to add support for Chef 12.1-12.4
+- Use apt_update resource vs. the apt cookbook in the test cookbook
+- Update Travis CI testing to use our standard Rakefile and cookstyle for ruby linting.
+- Fix chefspec / foodcritic / test kitchen failures
+- Swap the Policyfile for a Berksfile
+- Remove unnecessary action and default_action properties from the custom resources
+
+# v0.19.0
+
+- Remove delivery-cli examples and tests (we now shipit with ChefDK)
+- Set version constraint to ~> 1.1 for installing mixlib-install from Rubygems
+
+# v0.18.5
+
+- [#106](https://github.com/chef-cookbooks/chef-ingredient/pull/106) Limit `remote_file` backups to 1
+- [#110](https://github.com/chef-cookbooks/chef-ingredient/pull/110) Get rid of default: nil warnings
+
+# v0.18.4
+
+- Add `platform_version_compatibility_mode` property to `chef_ingredient` which makes chef-ingredient select packages built for earlier version of a platform during install when a package does not exist for the current platform version.
+
+# v0.18.3
+
+- Add `accept_license` property to `chef_ingredient` which can accept license for Chef products when applicable.
+
+# v0.18.2
+
+- Set version constraint to ~> 1.0 for installing mixlib-install from Rubygems
+
+# v0.18.1
+
+- Bump mixlib-install version to 1.0.6 so unstable channel artifacts won't include metadata.json files.
+
+# v0.18.0
+
+- [#85](https://github.com/chef-cookbooks/chef-ingredient/pull/85) Ability to support unstable channel for all products / platforms.
+- [#90](https://github.com/chef-cookbooks/chef-ingredient/pull/90) Use packages from packages.chef.io instead of package cloud & remove packagecloud repository setup.
+- [#91](https://github.com/chef-cookbooks/chef-ingredient/pull/91) Deprecate chef-ha, chef-marketplace, chef-sync, push-client, push-server in favor of ha, marketplace, sync, push-jobs-client, push-jobs-server.
+
+# v0.17.0
+
+- [#77](https://github.com/chef-cookbooks/chef-ingredient/pull/77) Enable installation of chef and chefdk from unstable
+- [#82](https://github.com/chef-cookbooks/chef-ingredient/pull/82) Set `--force-yes` for older Debian/Ubuntu
+
+# v0.16.0
+
+- [#62](https://github.com/chef-cookbooks/chef-ingredient/issues/62) Do not assume connection to the internet, allow custom recipe for a local repository
+- [#75](https://github.com/chef-cookbooks/chef-ingredient/pull/75) omnitruck handler windows implementation
+
+# v0.15.0
+
+- [#66](https://github.com/chef-cookbooks/chef-ingredient/pull/66) Fix push job client and server naming
+- [#68](https://github.com/chef-cookbooks/chef-ingredient/pull/68) Use mixlib-install while installing / upgrading packages from omnitruck
+- [#71](https://github.com/chef-cookbooks/chef-ingredient/pull/71) Convert `omnibus_service` and `ingredient_config` to "12.5 [custom resources](https://docs.chef.io/custom_resources.html)"
+- [#73](https://github.com/chef-cookbooks/chef-ingredient/pull/73) Use PRODUCT_MATRIX from mixlib-install
+
+# v0.14.0
+
+- [#58](https://github.com/chef-cookbooks/chef-ingredient/pull/58) Add Chef Compliance product
+
+
+# v0.13.1
+
+- [#57](https://github.com/chef-cookbooks/chef-ingredient/pull/57) Content accumulator guard
+
+# v0.13.0
+
+- [#56](https://github.com/chef-cookbooks/chef-ingredient/pull/56) Uncomment `use_inline_resources`, this is required for the providers to work properly
+- [#55](https://github.com/chef-cookbooks/chef-ingredient/pull/55) Remove unit tests for specifically the custom resources
+- [#54](https://github.com/chef-cookbooks/chef-ingredient/pull/54) Clarify maintainer/support in the README
+
+# v0.12.1
+
+- [#53](https://github.com/chef-cookbooks/chef-ingredient/pull/53) Relax version constraints
+
+# v0.12.0
+
+- Refactor `chef_ingredient` and prepare to handle install/upgrade from omnitruck
+- Add channel property to `chef_ingredient`
+- Removed installed state property
+- Use `product_name` instead of `package_name`
+- Add not if to skip `ingredient_config` render if `config` property isn't used
+
 # v0.11.3
 
 - Remove `resource_name` from Provider because `:facepalm:`